Updates PHP and Swagger
Why is this change needed?
--------------------------
Our versions of PHP and Swagger are out of date.

How does it address the issue?
------------------------------
This updates Swagger to version 2.4.14 and regenerates all of the agent
code. As a result we are also able to update our version of PHP to
version 7.2.

There are some new dependencies and API changes in the newly generated
code. Because of this the tests needed updated to auto load the vendored
libraries built by composer. The tests also needed to be updated to
utilize the new API syntax for creating and using the Configuration
class.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------
https://github.com/swagger-api/swagger-codegen/releases/tag/v2.4.14