Adds hosted documents to swagger config and regenerates code
Why is this change needed?
--------------------------
We need to add all of the new hosted document functionality and routes
to our PHP Agent.

How does it address the issue?
------------------------------
By updating the swagger configuration to include the hosted documents
routes we can then regenerate the code to include methods to help with
hosted document creation and expiration.

Previously, async docs were the only route that return a document status
object, but with the hosted document changes this is no longer true. The
AsyncDocStatus model has been rename to DocStatus so that it can be
shared between async documents and hosted documents.