Removes autoload.php and updates README
Why is this change needed?
--------------------------
Since version 2.3 swagger codegen requires a new dependency for making
HTTP requests. Because of these changes it is now more difficult to
manually install and require all of things needed to use this package.

Moving forward we will assume consumers of this package are using
composer to manage their dependencies.

How does it address the issue?
------------------------------
This removes the autoload.php file that no longer works and updates the
README to no longer include instructions about how to manually install
the package.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------
https://github.com/swagger-api/swagger-codegen/pull/5190