Adds examples for the new hosted document functionality
Why is this change needed?
--------------------------
The way hosted documents work is different that how DocRaptor has worked
in the past. Since we've added a new methods to help create hosted
documents we should add an example to portray how that works.

How does it address the issue?
------------------------------
This add new examples that create and host documents. After creating
the documents the examples output the public download URL. The file is
then downloaded and written to a temp directory to show the generated
document.

This also updates the other examples to include the API changes made to
the configuration model.