Updates the composer file to include our own information
Why is this change needed?
--------------------------
Swagger overwrites the composer file for us automatically every time we
regenerate the code. The composer file is used to provide information
about the package when we distribute it. We should update this to suit
our specific needs.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------
https://getcomposer.org/doc/04-schema.md#json-schema