Forces script/setup to install and link php@7.2
Why is this change needed?
--------------------------
Our setup script needs to make sure that we are using the right version
of PHP.

How does it address the issue?
------------------------------
There isn't a great way to manage multiple php versions. What we can do
is install PHP v7.2 with homebrew if it is not already installed.

php@7.2 is keg-only, which means it was not symlinked into /usr/local,
because this is an alternate version of another formula. Because of
this, we have to force it to be symlinked.