Updates tests and example to use the configuration created by the API class
Why is this change needed?
--------------------------
Currently we create an API configuration and pass it into the DocApi
class. The problem is that PHP doesn't have named arguments so we have
to set the first argument to null and the pass the configuration in as
the second argument. Which is gross and relies on our consumers
understanding which position the configuration is in.

How does it address the issue?
------------------------------
If we don't pass a configuration in to the DocApi when it is
instantiated it will make a new on by default. We can then get the
configuration and set the user name through that. This allows us to
avoid the confusion with positional arguments.

Additionally, the examples and tests did not use a relative path to find
resources. We've updated all of the files to require them using a path
relative to the directory that they are in.

Co-authored-by: David Ronk <david@expectedbehavior.com>