Updates the travis config
Why is this change needed?
--------------------------
Swagger stomps on our travis config when we regenerate the code.

How does it address the issue?
------------------------------
This adds .travis.yml to the .swagger-codegen-ignore file, preventing it
from being overwritten when we regenerate the code.The Travis config
file has also been update to run our test script rather than using
phpunit.

The composer file also now requires PHP 5.5 and above so we removed v5.4
from the list of versions that travis will test against.