Updates setup script
Why is this change needed?
--------------------------
There was a couple of issues with the setup script that would cause
problems when setting up the project for the first time.

How does it address the issue?
------------------------------
We now unlink the current php package if it already exists. This ensures
that the right version of php is linked to the `php` command.

This also tells composer to install all the required packages. We need
this to happen to utilize them later when we run the test or example
files.