Update script/setup to use homebrew php latest
Why is this change needed?
--------------------------
PHP 7.2 has been EOL for a while and 7.4 exits active support in 18 days[1].

How does it address the issue?
------------------------------
Switch to using homebrew's `php` formula so we're on 8.0. I'd prefer
to peg it to 8.0 (e.g. brew install php@8.0), but that doesn't seem to
be an option right now. I tried to get the asdf plugin working, but
the compilation failed even with all of the specified dependencies. I
could probably work out our own method of compiling it from scratch,
but I don't love the maintenance profile of that.

I'm going to come back in a bit and add something that does version
enforcement.

I also brought over a simple function for installing packages from
brew so the file looks a little cleaner.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------
1 - https://www.php.net/supported-versions.php