Use docker for testing
Why is this change needed?
--------------------------
In my previous commit, I updated the PHP version and promised to update the process to ensure the tests were consistently run against specific versions of PHP.

How does it address the issue?
------------------------------
It turned out to be extremely challenging to do this without docker, so I'm doing it with Docker. I thought I could just use asdf because that usually works for other languages, but the maintainers of asdf-php[1] have this to say

> To be honest, supporting a major version other than the latest without any extra work from the user is an endless endeavor that won't ever really work too well. It's not that we don't support them at all, but it's almost impossible for us to support them.

I tried anyway and there was "extra work" involved in the process. Instead of inflicting that extra work on you, let's test specific versions of PHP by downloading Docker images and running the tests inside that context.

That simplifies the setup script quite a bit. Instead of installing the docker images we need during setup, we'll just let docker fetch the images at execution time.

The `script/test` as you previously knew it has been moved to `script/inside_container/test` and is intended to only be executed inside containers with meaningful environments. That leaves `script/test` as a helper that executes the docker command on your behalf.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------
1 - https://github.com/asdf-community/asdf-php