Release v3.0.0
Why is this change needed?
--------------------------
PHP users want to support for Guzzle 7[1]

How does it address the issue?
------------------------------
OpenApi added support for Guzzle 7 as of OpenApi 5.3.0. At a high level, this commit

- Changes from Swagger to OpenApi
- Regenerates the client using OpenApi 5.3.0
- Adds stronger verification in the tests that generated documents are
  valid and correct

**BIG CAVEAT**: This client has been hand-edited in one important way.

The DocApi functions `getAsyncDoc` and `createDoc` have always
returned the file contents when called. Under OpenApi 5.3.0, the
generated client instead returns an SplFileObject[2]. If you attempt
to write an SplFileObject as if it werre the file contents, you get
something that looks and acts like a blank PDF. That's a very
confusing experience that could easily lead users to believe there are
problems with the service or the agent. There are lots of good reasons
to return an SplFileObject instead of the file contents, but in this
case it doesn't seem worth it. In the best case, it means everyone
that upgrades will need to make at least some changes to their
integration. At worst, it creates a very confusing and frustrating
failure.

If you look at those two functions, you'll see where I've commented
out the generated return and replaced it with the code necessary to
maintain the current behavior. Since I made that change by hand, it
will be destroyed when regenerated. If you're reading this because
you've regenerated the client and the tests are failing, compare the
code in the aforementioned DocApi functions from this commit to your
newly generated code and make the appropriate changes. In this case,
it is one line of fairly simple code. It should not be onerous to copy
or integrate. Or, I guess you could decide SplFileObject is the way to
go. The power is yours!

In other news, here's a high level of groups of change.

> .generator-revision, .openapi-generator-ignore, .openapi-generator

These files are renamed because we switched generators.

> lib/*

Files generated by OpenApi

> script/*

Mechanical changes to support OpenApi instead of Swagger. Things like
flag changes and referencing slightly different files that do
basically the same thing they've always done.

> test/*

Each test that makes a document now also verifies that the document is
a believable size. The real PDFs should be close to 30k. The "blank"
pdf problem I mentioned above generated documents less than 1k. If we
need a more robust test than this, we'll likely need to actually start
decoding the PDF.

Each test also writes the generated document to the local tmp
directory so it's easier to inspect when you think something hinky is
going on.

I also refactored the tests a bit to make it easier to share and
compare code. Hardcoded uses of the test name are replaced with a
variable and the code for writing the generated file to the file
system is now the same across each test.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------
1 - https://github.com/DocRaptor/docraptor-php/pull/24
2 - https://www.php.net/manual/en/class.splfileobject.php