Update browser download example with new code
Why is this change needed?
--------------------------
We made several improvements to the generated code and standardized
on that. However, the browser download example is not currently part
of the code generator.

How does it address the issue?
------------------------------
I manually updated the example so it matches the existing examples.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------