Switch to Dockerized openapi-generator
Why is this change needed?
--------------------------
I had an issue with my java configuration. It turned out to be unrelated
to this particular repo, but also we'd like for this repo to work
regardless of the environment.

Also this is much faster to setup than the previous way since it avoids
building the openapi-generator tools.

How does it address the issue?
------------------------------
Use a Dockerized package that behaves like the openapi-generator command
we're used to.