DRY up docker handling
Why is this change needed?
--------------------------
We want to regenerate all the clients in order to switch from using
the API through `docraptor.com` to `api.docraptor.com`. This allows us
to transition to a new load balancer that is more secure.

We'd like console and test to use the same docker command so dev and
test match as much as possible.

This also cleans up the language speicific bits so it's easier to
compare different language repos.

How does it address the issue?
------------------------------
Add a `script/docker` command that console and test use. Pull the
docker images into a separate file.

This also adds some prettier test output since the code was handy,
including test timing for a particular set of tests.

This includes some environment variables that'll get used in later
commits, but the file is necessary to run docker.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
https://3.basecamp.com/3093825/buckets/29124557/todos/5284412189
https://3.basecamp.com/3093825/buckets/29124557/todos/5284421778

Any screenshots?
----------------

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?