Improve script/setup
Why is this change needed?
--------------------------
Previously there was a worrysome warning. There could also be an error
of you try to `brew_install` a package where the formula and cask have
the same name.

How does it address the issue?
------------------------------
Switch the link behavior to avoid a warning.

Specify that we mean formulas when using `brew_install`.

Copy the nicer output from the ruby repo since it's easy and makes it
easier to compare contents across repos.

Clarify that the "good to go" output is coming from the setup script.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
https://3.basecamp.com/3093825/buckets/29124557/todos/5284412189
https://3.basecamp.com/3093825/buckets/29124557/todos/5284421778

Any screenshots?
----------------

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?