Allow a build script if necessary
Why is this change needed?
--------------------------
Some languages require a compilation process. It's nice if the test
script is the same between repos because then it's easier to compare
them for differences.

How does it address the issue?
------------------------------
If a build script exists we'll run it, otherwise we'll continue straight
to the tests.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
https://3.basecamp.com/3093825/buckets/29124557/todos/5284412189
https://3.basecamp.com/3093825/buckets/29124557/todos/5284421778

Any screenshots?
----------------

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?