Update openapi-generator-ignore files
Why is this change needed?
--------------------------
This will help us review changes with new versions of the agent code.

How does it address the issue?
------------------------------
Stop ignoring files where we want to review the autogenerated changes
and update related files.

Ideally there'd be a way to mark a new set of changes as having been
reviewed so we wouldn't have to deal with file changes every time we
generate the agent, but that is too complicated for the moment.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
https://3.basecamp.com/3093825/buckets/29124557/todos/5284421778
https://3.basecamp.com/3093825/buckets/29124557/todos/5284412189

Any screenshots?
----------------

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?