Commit composer.lock
Why is this change needed?
--------------------------
Mainly, this should increase the probability of the tests passing when
first pulling this repo and running the tests (because dependency
versions will be static).

Also this seems to be what's recommended in the community.

How does it address the issue?
------------------------------
Commit composer.lock and prevent the tests from altering it.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
https://3.basecamp.com/3093825/buckets/29124557/todos/5284412189
https://3.basecamp.com/3093825/buckets/29124557/todos/5284421778
https://getcomposer.org/doc/01-basic-usage.md#commit-your-composer-lock-file-to-version-control
http://getcomposer.org/doc/02-libraries.md#lock-file

Any screenshots?
----------------

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?