Add image specific cache dir
Why is this change needed?
--------------------------
Copied from the ruby repo, partially to keep everything as in sync as
possible and partly in case it's useful in the future to have an image
specific cache dir like it was in ruby.

How does it address the issue?
------------------------------
Update the docker script to mount an image specific cache dir.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
https://3.basecamp.com/3093825/buckets/29124557/todos/5284412189
https://3.basecamp.com/3093825/buckets/29124557/todos/5284421778

Any screenshots?
----------------

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?