Handle some quick shellcheck findings
Why is this change needed?
--------------------------
It's fast and good practice to run shellcheck when working on a bunch of
scripts.

How does it address the issue?
------------------------------
Run shellcheck and address reasonable/easy findings.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
https://3.basecamp.com/3093825/buckets/29124557/todos/5284412189
https://3.basecamp.com/3093825/buckets/29124557/todos/5284421778

Any screenshots?
----------------

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?