Fix test issue
Why is this change needed?
--------------------------
The check could return `false`, which is `==` 0 in PHP. This leads to an
incorrectly passing test when `%PDF` isn't found in the file content.

How does it address the issue?
------------------------------
Switch to `!==` which returns false if the operands are of different
types.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
https://3.basecamp.com/3093825/buckets/29124557/todos/5284412189
https://3.basecamp.com/3093825/buckets/29124557/todos/5284421778
https://www.php.net/manual/en/function.strpos.php
https://www.php.net/manual/en/language.operators.comparison.php

Any screenshots?
----------------

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?