Add generation dependency
Why is this change needed?
--------------------------
We use jq for some file processing, so it should be included in setup.

How does it address the issue?
------------------------------
Add `brew install jq` to setup.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------

Any screenshots?
----------------

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?