Add ENV var to indicate being inside docker
Why is this change needed?
--------------------------
Mostly to keep client code uniform. This can come in handy when writing
a script that can run outside the container or inside the container.

How does it address the issue?
------------------------------
Add an ENV var that gets applied in docker so we can check it to tell if
we are inside docker or not.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
https://3.basecamp.com/3093825/buckets/29007795/todolists/5257298600

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?