Standardize tracing across clients
Why is this change needed?
--------------------------
Tracing is helpful for debugging, so let's apply it consistently across
clients.

How does it address the issue?
------------------------------
Add a TRACE flag to the top of each script.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
https://3.basecamp.com/3093825/buckets/29007795/todolists/5257298600

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?